Suspend on low battery
Damn! This one took literally years to get right.

So the udev rules for thinkpad is a bit wonky, its not reported per percentage
decrease, but only when the /sys/class/power_supply/BAT1/energy_now value drops
below /sys/class/power_supply/BAT1/alarm.

Now this is kind of tricky because

1. I've no idea at what battery percentage this would actually happen. For now,
   it seems to be around 12%
2. Thinkpad has 2 batteries, BAT0 and BAT1 making the events even more
   unpredictable
3. Plugging in the cable will cause the same event to be fired,
   suspending the laptop.

But this seems to be a much better option than occasionally the machine dropping
dead and losing work :tada: :tada: